requirements.txt
fastapi
uvicorn
pyyaml
openai
anthropic
atws
